refactor : usecase 리팩토링 및 리뷰 수정사항 적용 (#207)
